chore: update submodule references for backend and frontend
This marks the end of building and maintaining of TBEPv1. Only very important bug fixes will be patched. We are now moving towards buidling TBEPv2 and this may some time and may bring many breaking changes in the middle so leaving this branch clean until we have some stable updates.